[svn r32696] changed dummy api method, missing from last patch applied

+++ modules/plugin/arcsde/sde-dummy/src/main/java/com/esri/sde/sdk/client/SeCoordinateReference.java
@@ -12,7 +12,7 @@ public class SeCoordinateReference {
 	public SeExtent getXYEnvelope() throws SeException { return null; }
 	public PeCoordinateSystem getCoordSys() { return null; }
 	
-	public void setXY(int i, int j, int k){}
+	public void setXY(double i, double j, double k){}
 	public void setXYByEnvelope(SeExtent s) {}
     public void setPrecision(int precision){}
     public void setCoordSysByID(SeObjectId seObjectId) {
